Move code into networking.go and util.go.
